backend/requirements.txt
# FastAPI وخادم الويب
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# قواعد البيانات والعملاء
psycopg2-binary==2.9.9
pymongo==4.6.0
influxdb-client==1.18.0
redis==5.0.1

# وسيط الرسائل Kafka
kafka-python==2.0.2

# معالجة البيانات والذكاء الاصطناعي
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# الأمان والمصادقة
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.1

# المراقبة والتسجيل
python-json-logger==2.0.7
prometheus-client==0.19.0

# أدوات التطوير
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0

# CORS والتكامل
aiohttp==3.9.1
httpx==0.25.2
